Using 32 bit particle iterators

+++ cosmictiger/defs.hpp
@@ -41,7 +41,7 @@
 
 #define NKICKS 25
 
-#define KICK_PP_MAX size_t(8*32)
+#define KICK_PP_MAX (8*32)
 #define MAX_BUCKET_SIZE 160
 
 #define MIN_ACTIVE_PER_BLOCK 16
